% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_markov_art.R
\name{make_markov_art}
\alias{make_markov_art}
\title{Make a Matrix of Values as a Basis for Random Artpiece Based on Very Simple Markov Chain}
\usage{
make_markov_art(size = 101, asp = 1, seed = NULL)
}
\arguments{
\item{size}{Integer value > 1 specifying the dimensions of the matrix to be traversed}

\item{asp}{If rectangular canvas is desired, integer value > 1 specifying the aspect ratio of output matrix.
This determines the number of columns by size * asp}

\item{seed}{Random seed for the Markov Chain}
}
\value{
A matrix of integer values
}
\description{
Make a Matrix of Values as a Basis for Random Artpiece Based on Very Simple Markov Chain
}
\examples{
make_markov_art(size = 11, asp = 1, seed = 42)
}
