% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiral_traversal.R
\name{spiral_traversal}
\alias{spiral_traversal}
\title{Get Order of Matrix-Cell-Coordinates From Inside Out Using Spiral Traversal Algorithm for Square Matrices}
\usage{
spiral_traversal(size)
}
\arguments{
\item{size}{Integer value specifying the dimensions of the matrix to be traversed}
}
\value{
A list with three elements, containing the coordinates of the matrix cells in
the spiral order as a 2 x (size^2) matrix; the number of rows as an integer value;
the number of columns as an integer value
}
\description{
Get Order of Matrix-Cell-Coordinates From Inside Out Using Spiral Traversal Algorithm for Square Matrices
}
\examples{
spiral_traversal(5)
}
