% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiral_traversal_rect.R
\name{spiral_traversal_rect}
\alias{spiral_traversal_rect}
\title{Get Order of Matrix-Cell-Coordinates From Inside Out Using Spiral Traversal Algorithm for Rectangular Matrices}
\usage{
spiral_traversal_rect(size, asp)
}
\arguments{
\item{size}{Integer value specifying the number of columns to be traversed}

\item{asp}{Integer value > 1 specifying the aspect ratio of output matrix.
This determines the number of columns by size * asp}
}
\value{
A list with three elements, containing the coordinates of the matrix cells in
the spiral order as a 2 x (size x (size x asp)) matrix; the number of rows as an integer value;
the number of columns as an integer value
}
\description{
Get Order of Matrix-Cell-Coordinates From Inside Out Using Spiral Traversal Algorithm for Rectangular Matrices
}
\examples{
spiral_traversal_rect(3, 2)
}
